+++ swap/src/bin/swap.rs
@@ -59,6 +59,7 @@ async fn main() -> Result<()> {
             seller,
             bitcoin_electrum_rpc_url,
             bitcoin_target_block,
+            bitcoin_change_address,
             monero_receive_address,
             monero_daemon_address,
             tor_socks5_port,
@@ -122,6 +123,7 @@ async fn main() -> Result<()> {
                 env_config,
                 event_loop_handle,
                 monero_receive_address,
+                bitcoin_change_address,
                 amount,
             );