+++ swap/src/protocol/bob.rs
@@ -33,10 +33,14 @@ impl Swap {
         env_config: env::Config,
         event_loop_handle: cli::EventLoopHandle,
         monero_receive_address: monero::Address,
+        bitcoin_change_address: bitcoin::Address,
         btc_amount: bitcoin::Amount,
     ) -> Self {
         Self {
-            state: BobState::Started { btc_amount },
+            state: BobState::Started {
+                btc_amount,
+                change_address: bitcoin_change_address,
+            },
             event_loop_handle,
             db,
             bitcoin_wallet,
@@ -47,6 +51,7 @@ impl Swap {
         }
     }
 
+    #[allow(clippy::too_many_arguments)]
     pub fn from_db(
         db: Database,
         id: Uuid,