+++ swap/tests/harness/mod.rs
@@ -434,6 +434,7 @@ impl BobParams {
             self.env_config,
             handle,
             self.monero_wallet.get_main_address(),
+            self.bitcoin_wallet.new_address().await?,
             btc_amount,
         );